refactor: slack, user - mq config 서로 바꿔서 설정되었던거 다시 설정#198
